+++ src/escape_from_darkness.h
@@ -3,11 +3,19 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdbool.h>
 #include <ncurses.h>
 #include <unistd.h>
+#include <ctype.h>
+#include <sys/stat.h>
+#include <sys/types.h>
 #include "dungeon.h"
+#include "settings.h"
+#include "resource.h"
 
-void initialize(Dungeon** dungeon);
-void terminate(Dungeon** dungeon);
+int initialize(Dungeon** dungeon, int argc, char* argv[]);
+int terminate(Dungeon** dungeon);
+void createApplicationFolders();
+int loadFunctionArguments(Settings* settings, int argc, char* argv[]);
 
 #endif